+++ libs/ngx-collection/src/lib/collection.service.spec.ts
@@ -1,6 +1,5 @@
 import { CollectionService } from './collection.service';
 import { of, Observable, timer, map } from 'rxjs';
-import { Comparator } from './comparator';
 
 beforeEach(() => {
   jest.useFakeTimers();
@@ -25,9 +24,7 @@ function readState<T>(stream: Observable<T>): T {
 }
 
 function setup() {
-  const coll = new CollectionService<Item>();
-  coll.setThrowOnDuplicates('duplicate');
-  coll.setComparator(new Comparator(['id']));
+  const coll = new CollectionService<Item>({throwOnDuplicates: 'duplicate', comparatorFields: ['id']});
   const vm = coll.getViewModel();
 
   return {
@@ -415,7 +412,8 @@ describe('Collection Service', () => {
     expect(readState(vm)).toMatchObject<ViewModel>({isUpdating: false, isProcessing: false, isMutating: false});
 
     coll.read({
-      request: of([item1, item2])
+      request: of([item1, item2]),
+      onSuccess: () => {throw 'oops';}
     }).subscribe();
 
     const newItem2 = {id: 0, name: 'C'};
@@ -714,8 +712,9 @@ describe('Collection Service', () => {
       }
     }
 
-    const dColl = new DColl();
-    dColl.setComparator((a, b) => a === b);
+    const dColl = new DColl({
+      comparator: (a, b) => a === b
+    });
     expect(dColl.hasDuplicates(
       [0, 1, 2, 3, 4, 5, 2]
     )).toStrictEqual(2);
@@ -734,8 +733,7 @@ describe('Collection Service', () => {
   });
 
   it('custom comparator fn with multiple fields', () => {
-    const coll = new CollectionService<Item>();
-    coll.setOptions({
+    const coll = new CollectionService<Item>({
       comparator: (item1: Item, item2: Item) => (item1.id + item1.name) === (item2.id + item2.name),
       allowFetchedDuplicates: false,
       onDuplicateErrCallbackParam: 'DRY'
@@ -764,4 +762,3 @@ describe('Collection Service', () => {
     );
   });
 });
-